# frontends/nextjs/todo-app/ts-api/models/TodoItemModel.ts
/* tslint:disable */
/* eslint-disable */
/**
 * TodoAPI
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface TodoItemModel
 */
export interface TodoItemModel {
    /**
     * 
     * @type {number}
     * @memberof TodoItemModel
     */
    id?: number;
    /**
     * 
     * @type {string}
     * @memberof TodoItemModel
     */
    name?: string;
    /**
     * 
     * @type {boolean}
     * @memberof TodoItemModel
     */
    isComplete?: boolean;
}

/**
 * Check if a given object implements the TodoItemModel interface.
 */
export function instanceOfTodoItemModel(value: object): boolean {
    return true;
}

export function TodoItemModelFromJSON(json: any): TodoItemModel {
    return TodoItemModelFromJSONTyped(json, false);
}

export function TodoItemModelFromJSONTyped(json: any, ignoreDiscriminator: boolean): TodoItemModel {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'] == null ? undefined : json['id'],
        'name': json['name'] == null ? undefined : json['name'],
        'isComplete': json['isComplete'] == null ? undefined : json['isComplete'],
    };
}

export function TodoItemModelToJSON(value?: TodoItemModel | null): any {
    if (value == null) {
        return value;
    }
    return {
        
        'id': value['id'],
        'name': value['name'],
        'isComplete': value['isComplete'],
    };
}